fix: 전체 scss 수정(북마크 크기, 글씨 크기등)
