+++ client/src/App.scss
@@ -3,7 +3,7 @@
 $black: #424242;
 $deep-blue: #5d83b1;
 $light-blue: #d8e5f4;
-$light-grey: #eeeeee;
+$light-grey: #e8e8e8;
 $deep-grey: #a0a0a0;
 $orange: #ebba9b;
 $white: #fafafa;
@@ -19,8 +19,8 @@ $f-deep-blue: invert(52%) sepia(70%) saturate(304%) hue-rotate(173deg)
   brightness(82%) contrast(85%);
 $f-light-blue: invert(86%) sepia(39%) saturate(571%) hue-rotate(184deg)
   brightness(109%) contrast(91%);
-$f-light-grey: invert(100%) sepia(1%) saturate(6628%) hue-rotate(293deg)
-  brightness(118%) contrast(87%);
+$f-light-grey: invert(100%) sepia(5%) saturate(68%) hue-rotate(207deg)
+  brightness(115%) contrast(82%);
 $f-deep-grey: invert(78%) sepia(7%) saturate(5%) hue-rotate(331deg)
   brightness(81%) contrast(89%);
 $f-orange: invert(81%) sepia(10%) saturate(1117%) hue-rotate(336deg)
