+++ client/src/styles/Mypage/Mypage.scss
@@ -1,5 +1,15 @@
 @import '../../App';
 
+.Mypage_topButton {
+  display: none;
+}
+
+@include tablet {
+  .Mypage_topButton {
+    display: block;
+  }
+}
+
 @include laptop {
   #mypage_container {
     display: flex;
