+++ client/src/styles/ReviewList.scss
@@ -85,10 +85,10 @@
 
 .ReviewList_list_none {
   width: 100%;
-  padding: 30px 30px 10px 30px;
-  // display: flex;
-  // justify-content: center;
-  // align-items: center;
+  height: 300px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 
   .review_none {
     width: 100%;
